+++ R/base.R
@@ -35,9 +35,9 @@
 
 # secretbase - Main Functions --------------------------------------------------
 
-#' Cryptographic Hashing Using the SHA-3 Algorithms
+#' SHA-3 Cryptographic Hash Algorithms and SHAKE256 XOF
 #'
-#' Returns a SHA-3 hash of the supplied object or file.
+#' Returns a SHA-3 or SHAKE256 hash of the supplied object or file.
 #'
 #' @param x object to hash. A character string or raw vector (without
 #'     attributes) is hashed 'as is'. All other objects are stream hashed using
@@ -99,9 +99,9 @@ sha3 <- function(x, bits = 256L, convert = TRUE, file)
   if (missing(file)) .Call(secretbase_sha3, x, bits, convert) else
     .Call(secretbase_sha3_file, file, bits, convert)
 
-#' Cryptographic Hashing Using the SHA-256 Algorithm
+#' SHA-256 Cryptographic Hash Algorithm
 #'
-#' Returns a SHA-256 hash of the supplied object or file, or an HMAC if a secret
+#' Returns a SHA-256 hash of the supplied object or file, or HMAC if a secret
 #'     key is supplied.
 #'
 #' @inheritParams sha3
@@ -144,10 +144,12 @@ sha256 <- function(x, key = NULL, convert = TRUE, file)
   if (missing(file)) .Call(secretbase_sha256, x, key, convert) else
     .Call(secretbase_sha256_file, file, key, convert)
 
-#' Hashing Using the SipHash-1-3 Pseudorandom Function
+#' SipHash Pseudorandom Function
 #'
-#' Returns a fast, cryptographically-strong SipHash-1-3 keyed hash of the
-#'     supplied object or file.
+#' Returns a fast, cryptographically-strong SipHash keyed hash of the supplied
+#'     object or file. SipHash-1-3 is optimised for performance, whereas
+#'     SipHash-2-4 is recommended for security. Note: SipHash is not a
+#'     cryptographic hash algorithm.
 #'
 #' @inheritParams sha3
 #' @param key [default NULL] a character string or raw vector comprising the 16
@@ -177,11 +179,11 @@ sha256 <- function(x, key = NULL, convert = TRUE, file)
 #' # SipHash-1-3 hash as raw vector:
 #' siphash13("secret base", convert = FALSE)
 #' 
-#' # SipHash-1-3 hash using a character string key:
-#' siphash13("secret", key = "base")
+#' # SipHash-2-4 hash using a character string key:
+#' siphash24("secret", key = "base")
 #' 
-#' # SipHash-1-3 hash using a raw vector key:
-#' siphash13("secret", key = charToRaw("base"))
+#' # SipHash-2-4 hash using a raw vector key:
+#' siphash24("secret", key = charToRaw("base"))
 #' 
 #' # SipHash-1-3 hash a file:
 #' file <- tempfile(); cat("secret base", file = file)
@@ -193,3 +195,10 @@ sha256 <- function(x, key = NULL, convert = TRUE, file)
 siphash13 <- function(x, key = NULL, convert = TRUE, file)
   if (missing(file)) .Call(secretbase_siphash13, x, key, convert) else
     .Call(secretbase_siphash13_file, file, key, convert)
+
+#' @rdname siphash13
+#' @export
+#'
+siphash24 <- function(x, key = NULL, convert = TRUE, file)
+  if (missing(file)) .Call(secretbase_siphash24, x, key, convert) else
+    .Call(secretbase_siphash24_file, file, key, convert)