+++ NEWS.md
@@ -1,7 +1,7 @@
-# secretbase 0.3.0.9006 (development)
+# secretbase 0.3.0.9007 (development)
 
 * Adds HMAC generation to `sha256()`.
-* Adds SipHash-1-3 pseudo-random function (PRF) as a fast, cryptographically-strong keyed hash.
+* Adds SipHash pseudo-random function (PRF) as a fast, cryptographically-strong keyed hash.
 
 # secretbase 0.3.0.1